+++ CHANGES.md
@@ -1,3 +1,12 @@
+## Version 8.5.0
+
+Unreleased
+
+- Supported versions of Windows enable ANSI terminal styles by default.
+  Colorama is no longer a dependency and is not used. {issue}`2986` {pr}`3505`
+- {class}`Argument` accepts a `help` parameter, and help output includes
+  a `Positional arguments` section when argument help is available. {issue}`2983` {pr}`3473`
+
 ## Version 8.4.2
 
 Released 2026-06-24